requirements_alt.txt
annotated-types==0.6.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
dnspython==2.6.1
exrex==0.11.0
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.43.1
geneticalgorithm2==6.8.5
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
knapsack==0.1.1
langcodes==3.3.0
MarkupSafe==2.1.5
matplotlib==3.8.0
more-itertools==10.1.0
murmurhash==1.0.10
numpy==1.26.1
OppOpPopInit==2.0.1
ortoolpy==0.6.4
packaging==23.2
pandas==2.1.1
Pillow==10.0.1
preshed==3.0.9
PuLP==2.7.0
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.7.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
setuptools==69.1.1
six==1.16.0
smart-open==6.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stem==1.8.2
thinc==8.2.3
tqdm==4.66.2
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
